% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{Forest_Tester_AD}
\alias{Forest_Tester_AD}
\title{Anderson darling tester with argument for parameter.}
\usage{
Forest_Tester_AD(params, dataset)
}
\arguments{
\item{params}{data.frame with stable distribution parameters}

\item{dataset}{data.frame a data frame with colour intensities corresponding
to red, green and blue channels of the input image.}
}
\value{
a data frame with statistics of AD tests corresponding to red, green
  and blue channels of the input image.
}
\description{
Anderson darling tester with argument for parameter.
}
\examples{
library(deforeStable)
data("geoimages")

obj <- geoimages[[2]]
plotRGB(obj, scale=1, asp=1)

mtrx <- as.matrix(obj)
pars <- Koutparams(mtrx)

dd_ad<-Forest_Tester_AD(params=pars, dataset=mtrx)
}
