% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{Forest_Tester_CVM}
\alias{Forest_Tester_CVM}
\title{Cramer von Mises tester with argument for parameter}
\usage{
Forest_Tester_CVM(params, dataset)
}
\arguments{
\item{params}{data.frame with stable distribution parameters}

\item{dataset}{data.frame a data frame with colour intensities corresponding
to red, green and blue channels of the input image.}
}
\value{
a data frame with statistics of CvM tests corresponding to red, green
  and blue channels of the input image.
}
\description{
Performs Cramer-von Mises goodness-of-fit tests of empirical distributions of
colour intensities in the three channels against pre-specified stable
distributions. It is designed to be used in conjunction with Koutparams, and
argument params is a data frame returned by the latter. dataset is a data
frame with colour intensities of the image. It has the same format as the one
produced by read_data: observations are in rows and each of the 3 columns
correspond to red, green or blue colour.
}
\examples{

library(deforeStable)
data("geoimages")

obj <- geoimages[[2]]
plotRGB(obj, scale=1, asp=1)

mtrx <- as.matrix(obj)
pars <- Koutparams(mtrx)

dd_cvm <- Forest_Tester_CVM(params=pars, dataset=mtrx)
dd_cvm
}
