% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{Mahala_dist}
\alias{Mahala_dist}
\title{Squared Mahalanobis distance between two samples}
\usage{
Mahala_dist(params, dataset)
}
\arguments{
\item{params}{data.frame with colour intensities corresponding to red, green
and blue channels of the input image.}

\item{dataset}{data.frame with colour intensities corresponding to red, green
and blue channels of the input image.}
}
\value{
a numeric with value of squared Mahalanobis distance.
}
\description{
Calculates the squared Mahalanobis distance between two distributions of
colour intensities. Argument params is a data frame with the reference
disitribution to test against and argument dataset is a data frame with the
distribution to test. Both have the same format as the one produced by
read_data: observations are in rows and each of the 3 columns correspond to
red, green or blue colour.
}
\examples{

library(deforeStable)
library(raster)
data("geoimages")

obj1 <- geoimages[[2]]
obj2 <- geoimages[[11]]
plotRGB(obj1, scale=1, asp=1)
plotRGB(obj2, scale=1, asp=1)

mtrx <- as.matrix(obj1)
pars <- as.matrix(obj2)

dd_Sqmahala <- Mahala_dist(params=pars, dataset=mtrx)
dd_Sqmahala
}
