% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{mahalaclust}
\alias{mahalaclust}
\title{Clusters images of forest}
\usage{
mahalaclust(ForestData, clusters, Forest_params)
}
\arguments{
\item{ForestData}{list of named matrices containing data of forest image
color intensities, as returned by read_data.}

\item{clusters}{a numeric giving the number of clusters to divide the forest
data into}

\item{Forest_params}{list of named data frames with stable distribution
parameters, as retuned by function Koutparams.}
}
\value{
list of data frames with averaged stable distribution parameter set
  for each cluster.
}
\description{
Clusters images of forest using hierarchical clustering and the squared
Mahalanobis distance in order to reduce the number of parameter sets used for
testing in the parametric model implemented in Param_classifier. Images that
are clustered together are assumed to be from the same distribution and have
their estimated stable distribution parameters averaged and returned. Inputs
are a list of forest image data on matrix form, the number of clusters to
cluster the data into and a list of data frames containing the estimated
stable distribution parameters for the provided forest images.
}
\examples{
library(deforeStable)
library(raster)
data("geoimages")
data("geoimages_desc")

data <- geoimages[1:5]

datalist <- lapply(data, as.matrix)
parlist <- lapply(datalist, Koutparams)
names(datalist) <- as.list(geoimages_desc[1:5,1])
names(parlist) <- as.list(geoimages_desc[1:5,1])

clustpars <- mahalaclust(ForestData = datalist, clusters = 2, Forest_params = parlist)
clustpars
}
