% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{MultipleTester}
\alias{MultipleTester}
\title{Do multiple parametric tests and return best one}
\usage{
MultipleTester(data, params, func)
}
\arguments{
\item{data}{a data frame with colour intensities corresponding to red, green
and blue channels of the input image.}

\item{params}{a list of data frames with parameter values}

\item{func}{a function that performs a test and returns its statistic}
}
\value{
data frame with smallest statistic.
}
\description{
Performs multiple hypothesis tests according to function func of empirical
data data against a list of parameters params. func and params must be
compatible, apart from that there are no specific restrictions imposed on
them.
}
\examples{
library(deforeStable)
library(raster)
data("geoimages")
obj <- geoimages[[10]]
plotRGB(obj, scale=1, asp=1)

mtrx <- as.matrix(obj)
pars <- Koutparams(mtrx)

mpt <- MultipleTester(data=mtrx, params=list(pars, pars), func=Forest_Tester_CVM)
mpt
}
