% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{NonParamCV}
\alias{NonParamCV}
\title{Cross-validation for the Non-parametric model using the squared Mahalanobis
distance}
\usage{
NonParamCV(forestdir, Nonforestdir, n_pts = 7, nrfolds = 5)
}
\arguments{
\item{forestdir}{character string with path to directory containing forest
images.}

\item{Nonforestdir}{character string with path to directory containing
non-forest images.}

\item{n_pts}{numeric for sub-image side length, default=7.}

\item{nrfolds}{numeric for number of folds in k-fold cross-validation,
default=5.}
}
\value{
list with first element: numeric with best accuracy and threshold
  which produced it. Second element: data frame with all tested thresholds
  and corresponding accuracy and other performance metrics. Third element:
  List of data frames with color intensities of forest images used to train.
}
\description{
Performs k-fold cross-validation to find the optimal threshold for the
squared Mahalanobis distance to use for the non parametric model implemented
in function Nonparam_classifier. This is done by splitting images into
smaller sub-images and and for each sub-image the squared Mahalanobis
distance is computed with all forest images in the data set and the smallest
distance is chosen. These distances are compared to a range of thresholds and
the threshold which produces the best classification result according to the
accuracy is returned. As input it takes a path to a directory containing
forest images, a path to a directory contain non-forest images, the side
length of the sub-images the images are split into and the number of folds to
use for the k-fold cross-validation. Computations are parallelized.
}
\examples{

library(deforeStable)
library(doParallel)

forestdir <- "forest image directory path"
Nonforestdir <- "Non-forest image directory path"
NonparCV <- NonParamCV(forestdir = forestdir, Nonforestdir = Nonforestdir,
                       n_pts = 7, nrfolds = 5)

test_image <- read_data_raster(filename, dir)

Nonpar_pred <- Nonparam_classifier(test_image, n_pts = 7, references = NonparCV[[3]],
                          thres = NonparCV[[1]][[1]])
jpeg::writeJPEG(image=Nonpar_pred, target='Nonpartest_im.jpeg')



}
