% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{Param_classifier}
\alias{Param_classifier}
\title{Classification of an image using the parametric model based on the Cramér-von
Mises statistics}
\usage{
Param_classifier(rastData, n_pts, pars, thres, parallel = TRUE)
}
\arguments{
\item{rastData}{RasterStack as returned by function read_data_raster.}

\item{n_pts}{numeric for sub-image side length.}

\item{thres}{numeric thresholds for Cramér-von Mises statistics}

\item{parallel}{logical for activating parallel computing.}

\item{references}{list of data frames with stable distribution parameters, as
produced by function Koutparams.}
}
\value{
A matrix of same dimension as input image, with element values 0 or
  1. Element value 0 indicates that pixel does not contain reference terrain,
  1 indicates the opposite.
}
\description{
Predict whether parts of a given image contains the terrains of provided
reference stable distributions by splitting the given image into smaller
sub-images and testing these against reference disitributions. The argument
rastData is the image to be classified on the RasterStack format. Argument
n_pts is the side length of square sub-images the image is split into.
Argument pars is a list of data frames with parameters for the stable
distributions of colour intensities of the reference images. Argument thres
is threshold values for the squared Cramér-von Mises statistics, deciding if
a sub-image belongs to reference terrain or not. Parameter sets and threshold
should be obtained from ParamCV function. Parallel decides if computations
should be parallelized.
}
\examples{

library(deforeStable)
library(raster)
library(doParallel)
data("geoimages")

obj <- geoimages[[28]]
ref <- geoimages[1:5]

reflist <- lapply(ref, as.matrix)
parlist <- lapply(reflist, Koutparams)

parClass <- Param_classifier(rastData = obj, n_pts = 7, pars = parlist, thres = c(15,15,15), parallel = TRUE)
jpeg::writeJPEG(image = parClass, target = 'parClass.jpeg')
}
