% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifiers_estimators.R
\name{ParamCV}
\alias{ParamCV}
\title{Cross-validation for the parametric model using the Cramér-von Mises
statistic and stable distributions.}
\usage{
ParamCV(
  forestdir,
  Nonforestdir,
  n_pts = 7,
  nrfolds = 5,
  clustering = TRUE,
  maxt = 14
)
}
\arguments{
\item{forestdir}{character string with path to directory containing forest
images.}

\item{Nonforestdir}{character string with path to directory containing
non-forest images.}

\item{n_pts}{numeric for sub-image side length, default=7.}

\item{nrfolds}{numeric for number of folds in k-fold cross-validation,
default=5.}

\item{clustering}{logical for deciding if clustering of forest image
parameter sets should be included, default=TRUE.}

\item{maxt}{numeric for the max threshold range that the CvM statistic is
checked against to maximize accuracy, default=14.}
}
\value{
list with first element: numeric with best accuracy and threshold
  which produced it. Second element: data frame with all tested thresholds
  and corresponding accuracy and other performance metrics. Third element:
  List of data frames with parameter sets for the stable distributions of the
  color intensities of forest images used to train.
}
\description{
Performs k-fold cross-validation to find the optimal threshold for the
Cramér-von Mises statistic to use for the parametric model implemented in
function Param_classifier. This is done by splitting images into smaller
sub-images and and for each sub-image the Cramér-von Mises statistics for
each of the three color intensities red, green and blue is computed with all
forest images in the data set and the smallest sum of statistics is chosen.
These statistics are compared to a range of thresholds and the threshold
which produces the best classification result according to the accuracy is
returned. As input it takes a path to a directory containing forest images, a
path to a directory contain non-forest images, the side length of the
sub-images the images are split into and the number of folds to use for the
k-fold cross-validation. Further, it takes a logical deciding if clustering
of parameter sets should be included, to reduce computational time. Lastly,
it takes the max threshold for the range checked against when trying to find
the optimal Cramér-von Mises threshold. Computations are parallelized.
}
\examples{

library(deforeStable)
library(doParallel)

forestdir <- "forest image directory path"
Nonforestdir <- "Non-forest image directory path"
ParCV <- ParamCV(forestdir = forestdir, Nonforestdir = Nonforestdir,
                    n_pts = 7, nrfolds = 5, clustering = TRUE,
                    maxt = 14)

test_image <- read_data_raster(filename, dir)

Par_pred <- Param_classifier(test_image, n_pts = 7, pars = ParCV[[3]],
                          thresh = ParCV[[1]][1:3])
jpeg::writeJPEG(image=Par_pred, target='Partest_im.jpeg')

}
