% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{read_data}
\alias{read_data}
\alias{read_data_matrix}
\alias{read_data_raster}
\title{Import a jpeg image}
\usage{
read_data(filename, dir)

read_data_matrix(filename, dir)

read_data_raster(filename, dir)
}
\arguments{
\item{filename}{name of the jpeg file to import}

\item{dir}{the directory where the image is located}
}
\description{
read_data is to read jpeg images and return a 3-column data.frame with pixels in rows and
red, green, blue in columns. read_data_matrix reads jpeg images and return 3 matrices for
each of red, green and blue colors. read_data_raster imports jpeg as a raster object.
}
\section{Functions}{
\itemize{
\item \code{read_data_matrix}: returns three matrices

\item \code{read_data_raster}: returns a RasterStack object
}}

\examples{
dd<-read_data(filename, dir)
hist(dd[,1])

dd<-read_data_matrix(filename, dir)
jpeg::writeJPEG(image=dd[[1]], target='ex.jpeg')

dd<-read_data_raster(filename, dir)
plotRGB(dd, scale=1, asp=1)
}
